requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
yt-dlp==2024.1.30
google-generativeai==0.3.2
python-multipart==0.0.6
